Bug #6548 【PC端】表扬功能的分数换算不正确。由于强转会舍去小数点后的数字，所以如果出现0.5不能进一而会被舍去

+++ teacher_app/src/cn/com/incito/interclass/ui/widget/PraiseDialog.java
@@ -318,7 +318,7 @@ public class PraiseDialog extends JDialog implements MouseListener {
 					if (jsonObject.getIntValue("code") == 1) {
 						return;
 					} else {
-						double scoreResult = (jsonObject.getIntValue("score") / group.getStudents().size());
+						double scoreResult = (jsonObject.getIntValue("score") / (float)group.getStudents().size());
 						for (int i = 0; i < group.getStudents().size(); i++) {
 							if (group.getStudents().get(i).isLogin()) {
 								group.getStudents().get(i).setScore(group.getStudents().get(i).getScore() + updateScore);
